small changes : posts verifies that you are logged in to like a post and button is now prettier in uploadGuitaristView.php

+++ posts.php
@@ -42,34 +42,36 @@ if ($browsePosts) {
 
 
     // ------------- Like interraction ------------- //
-    if(!empty($_POST['id_guit']) and !empty($_POST['like'])) {
-        $like = ($_POST['like'] === '⬆️') ? 1 : 0;
-        
-        //Remove from db
-        $sqlQuerry = "DELETE FROM appreciation WHERE id_user = :id_user && id_guitarist = :id_guitarist;";
-        
-        $statement = $db->prepare($sqlQuerry);
-        $statement->execute([
-            'id_user' => $_SESSION['id_user'],
-            'id_guitarist' => $_POST['id_guit']
-        ]);
-
-        // Add in db
-        if ($like === 1 and in_array($_POST['id_guit'], $likedPosts))/* Do nothing */;
-        else if($like === 0 and in_array($_POST['id_guit'], $dislikedPosts))/* Do nothing */;
-        else {
-            $sqlQuerry = "INSERT INTO appreciation (id_user, id_guitarist, likes) VALUES (:id_user, :id_guitarist, :likes);";
+    if(isset($_SESSION['id_user'])) {
+        if(!empty($_POST['id_guit']) and !empty($_POST['like'])) {
+            $like = ($_POST['like'] === '⬆️') ? 1 : 0;
+            
+            //Remove from db
+            $sqlQuerry = "DELETE FROM appreciation WHERE id_user = :id_user && id_guitarist = :id_guitarist;";
             
             $statement = $db->prepare($sqlQuerry);
             $statement->execute([
                 'id_user' => $_SESSION['id_user'],
-                'id_guitarist' => $_POST['id_guit'],
-                'likes' => $like
+                'id_guitarist' => $_POST['id_guit']
             ]);
 
+            // Add in db
+            if ($like === 1 and in_array($_POST['id_guit'], $likedPosts))/* Do nothing */;
+            else if($like === 0 and in_array($_POST['id_guit'], $dislikedPosts))/* Do nothing */;
+            else {
+                $sqlQuerry = "INSERT INTO appreciation (id_user, id_guitarist, likes) VALUES (:id_user, :id_guitarist, :likes);";
+                
+                $statement = $db->prepare($sqlQuerry);
+                $statement->execute([
+                    'id_user' => $_SESSION['id_user'],
+                    'id_guitarist' => $_POST['id_guit'],
+                    'likes' => $like
+                ]);
+
+            }
+            
+            header('Location: posts.php?page=' . $page . ((isset($_GET['sort'])) ? '&sort=' . $_GET['sort'] : ''));
         }
-        
-        header('Location: posts.php?page=' . $page . ((isset($_GET['sort'])) ? '&sort=' . $_GET['sort'] : ''));
     }
 
     include("postsViewBrowse.php");
